% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARUtoolsExtra-package.R
\docType{package}
\name{ARUtoolsExtra}
\alias{ARUtoolsExtra-package}
\alias{ARUtoolsExtra}
\title{ARUtoolsExtra: Management and Processing of Autonomous Recording Unit (ARU) Data}
\description{
Parse Autonomous Recording Unit (ARU) data and for sub-sampling recordings.
Extract Metadata from your recordings, select a subset of recordings for
interpretation, and prepare files for processing on the
WildTrax <https://wildtrax.ca/> platform. Read and process metadata
from recordings collected using the Song Meter and BAR-LT types of ARUs.
}
